add call-module and call-module-with-block
In OpenSCAD it is clear enough when you are calling a module with a
block: there are curly braces and child objects. But in Clojure that
block is just the last parameter. So, in order to be able to pass
parameters to an OpenSCAD module, we have to distinguish whether the
last one is a block or not. The simplest way to do that is to compel the
caller to specify.

The purpose for this is that, in my fork of the Dactyl keyboard, I
key-place a lot of things. But every time I say key-place in Clojure, it
results in a chain of seven or eight transforms being written to the
OpenSCAD output. This makes OpenSCAD take a great deal of time, CPU, and
RAM just to read the input, much less to show me the thing I've
described to it with such dull, interminable monotony. So I decided to
make key-place into an OpenSCAD module. But then I need to emit a call
to the module from my Clojure code. (At some point I may need to emit a
definition of the module, but I'm not there yet.)