add define-module
It is possible to pass argument names to define-module as well as a
body. No means is added by which you can use the named arguments in
expressions from Clojure. This makes define-module mostly useful for
defining a module with no parameters that describes a complicated shape.
Then you can talk about the shape again using call-module, and avoid
OpenSCAD output ballooning to megabytes.